perf(test): remove real-time waits and split CopyKeyDialog test file
- tempWindowPool turnstile timeout test: replace 2.5s real sleep with
  fake-timer advancement (matches other suites in the file)
- CopyKeyDialog copied-state reset: advance the component's 2s reset
  timer with fake timers instead of a real-time waitFor(2500)
- ShieldBypassPromptToast title-correction test: wait for correction
  convergence and assert observer disconnect deterministically instead
  of sleeping 1100ms of real time
- redemptionAssist self-click test: flush one macrotask instead of
  sleeping 700ms
- Split CopyKeyDialog.test.tsx (53 tests, ~100s CI wall time, largest
  single-file tail) into 4 files grouped by subject with shared mocks/
  support modules so vitest can parallelize them within a shard
